refactor: reorganize room types and update imports
