+++ src/types/chat.ts
@@ -1,0 +1,13 @@
+export enum ChatRoomKind {
+  All = 'all',
+  White = 'white',
+  Black = 'black',
+  User01 = 'user01',
+  User02 = 'user02',
+  User03 = 'user03',
+  User04 = 'user04',
+  User05 = 'user05',
+  User06 = 'user06',
+  User07 = 'user07',
+  User08 = 'user08',
+}
