--- src/types/form.ts
@@ -1,8 +1,0 @@
-export type CreateRoomForm = {
-  name: string;
-  password: string;
-};
-
-export type CreateRoomResponse = {
-  id: string;
-};
