--- src/types/game.ts
@@ -1,29 +1,0 @@
-import { type Username } from '~/types/account';
-
-export type Room = {
-  id: string;
-  name: string;
-  owner: Username;
-  maxUser: number;
-  currentUser: number;
-  isPlaying: boolean;
-};
-
-export type PaginatedRoomsResponse = {
-  rooms: Room[];
-  last: boolean;
-};
-
-export enum ChatRoomKind {
-  All = 'all',
-  White = 'white',
-  Black = 'black',
-  User01 = 'user01',
-  User02 = 'user02',
-  User03 = 'user03',
-  User04 = 'user04',
-  User05 = 'user05',
-  User06 = 'user06',
-  User07 = 'user07',
-  User08 = 'user08',
-}
